If uploading image och document with the same name and storage backend doesn't change the filename, don't delete the old file, becasue it will really be the new one.

+++ wagtail/documents/views/documents.py
@@ -195,10 +195,12 @@ def edit(request, document_id):
                 doc.save()
                 form.save_m2m()
 
-                # If providing a new document file, delete the old one.
+                # If providing a new document file, delete the old one, unless it has
+                # the same name. 3rd party storage backends might work that way
                 # NB Doing this via original_file.delete() clears the file field,
                 # which definitely isn't what we want...
-                original_file.storage.delete(original_file.name)
+                if original_file.name != doc.file.name:
+                    original_file.storage.delete(original_file.name)
             else:
                 doc = form.save()
